+++ NewBasicTemplate2016/src/Subsystems/Arm.h
@@ -77,7 +77,7 @@ public:
 	void SetEncoderPosition(int);
 	double GetPotValue();
 
-	double GetPotOrEncoderValueForAutomationOfArm(double);
+	int GetPotOrEncoderValueForAutomationOfArm(double);
 	double GetPotValueOrEncoderPosition();
 
 };
